refactor: minor fix
